Separate consume and compute routes.
