+++ ocean_provider/routes/__init__.py
@@ -1,0 +1,7 @@
+# flake8: noqa
+from flask import Blueprint
+
+from .compute import *  # isort:skip
+from .consume import *  # isort:skip
+
+services = Blueprint("services", __name__)
